Inital Scheduler querys
