+++ src/components/FeatureViews/Courses/CourseClass.js
@@ -25,6 +25,7 @@ import AccessControlComponent from '../../OmouComponents/AccessControlComponent'
 import AttendanceContainer from './AttendanceContainer';
 import { StudentCourseLabel } from './StudentBadge';
 import { GET_STUDENTS } from './CourseManagementContainer';
+import {GET_COURSE} from "../../../queries/CoursesQuery/CourseQuery"
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -73,82 +74,7 @@ export const GET_ANNOUNCEMENTS = gql`
     }
 `;
 
-export const GET_CLASS = gql`
-    query getClass($id: ID!) {
-        course(courseId: $id) {
-            academicLevel
 
-            title
-            startDate
-            endDate
-            description
-            courseLink
-            courseLinkDescription
-            courseLinkUpdatedAt
-            courseLinkUser {
-                firstName
-                lastName
-            }
-            courseCategory {
-                name
-                id
-            }
-            activeAvailabilityList {
-                dayOfWeek
-                endTime
-                startTime
-            }
-            instructor {
-                user {
-                    firstName
-                    lastName
-                }
-            }
-            enrollmentSet {
-                student {
-                    user {
-                        firstName
-                        lastName
-                        id
-                    }
-                    primaryParent {
-                        user {
-                            firstName
-                            lastName
-                            id
-                            email
-                        }
-                        accountType
-                        phoneNumber
-                    }
-                    studentschoolinfoSet {
-                        textbook
-                        teacher
-                        name
-                    }
-                    accountType
-                }
-            }
-            sessionSet {
-                startDatetime
-                id
-            }
-
-            availabilityList {
-                startTime
-                endTime
-                dayOfWeek
-            }
-        }
-        enrollments(courseId: $id) {
-            student {
-                user {
-                    id
-                }
-            }
-        }
-    }
-`;
 
 const CourseClass = () => {
     const { id } = useParams();
@@ -175,7 +101,7 @@ const CourseClass = () => {
 
     const parentNostudentEnrolledTab = [{ label: 'About Course' }];
 
-    const { data, loading, error } = useQuery(GET_CLASS, {
+    const { data, loading, error } = useQuery(GET_COURSE, {
         variables: {
             id: id,
         },
