+++ src/components/FeatureViews/Registration/CourseRegistrationReceipt.js
@@ -11,32 +11,12 @@ import Moment from 'react-moment';
 import Avatar from '@material-ui/core/Avatar';
 import { stringToColor } from '../Accounts/accountUtils';
 import Divider from '@material-ui/core/Divider';
-import Button from '@material-ui/core/Button';
 import NavLinkNoDup from '../../Routes/NavLinkNoDup';
 import CourseAvailabilites from '../../OmouComponents/CourseAvailabilities';
-
+import {GET_COURSE} from "../../../queries/CoursesQuery/CourseQuery";
 import { ResponsiveButton } from '../../../theme/ThemedComponents/Button/ResponsiveButton';
 
-const GET_COURSE = gql`
-    query GetCourse($courseId: ID!) {
-        course(courseId: $courseId) {
-            title
-            startDate
-            availabilityList {
-                endTime
-                startTime
-                dayOfWeek
-            }
-            endDate
-            instructor {
-                user {
-                    firstName
-                    lastName
-                }
-            }
-        }
-    }
-`;
+
 
 export default function CourseRegistrationReceipt({ formData, format }) {
     const { type } = useParams();
