+++ src/components/FeatureViews/Scheduler/EditSessionView.js
@@ -4,9 +4,7 @@ import Grid from '@material-ui/core/Grid';
 import {
     useHistory,
     useLocation,
-    withRouter,
     useParams,
-    NavLink,
 } from 'react-router-dom';
 import Loading from 'components/OmouComponents/Loading';
 
@@ -20,7 +18,6 @@ import * as userActions from '../../../actions/userActions.js';
 import { useDispatch, useSelector } from 'react-redux';
 import { FormControl, Typography } from '@material-ui/core';
 import * as apiActions from '../../../actions/apiActions';
-import Button from '@material-ui/core/Button';
 import TextField from '@material-ui/core/TextField';
 import { DatePicker } from '@material-ui/pickers/DatePicker/DatePicker';
 import { TimePicker } from '@material-ui/pickers/TimePicker/TimePicker';
@@ -31,7 +28,6 @@ import MenuItem from '@material-ui/core/MenuItem';
 
 import { dateFormat, timeFormat } from '../../../utils';
 import InstructorConflictCheck from 'components/OmouComponents/InstructorConflictCheck';
-import BackButton from '../../OmouComponents/BackButton';
 import { ResponsiveButton } from 'theme/ThemedComponents/Button/ResponsiveButton';
 import './scheduler.scss';
 import GET_SESSIONS from '../Accounts/TabComponents/EnrollmentView';
@@ -41,14 +37,8 @@ import { fullName } from '../../../utils';
 import moment from 'moment';
 import { GET_SESSIONS_FAILED } from 'actions/actionTypes';
 
-const GET_CATEGORIES = gql`
-    query EditSessionCategoriesQuery {
-        courseCategories {
-            id
-            name
-        }
-    }
-`;
+import {GET_CATEGORIES} from "../../../queries/CoursesQuery/CourseQuery"
+import {GET_SESSION} from "../../../queries/SchedulerQuery/SchedulerQuery"
 
 const GET_INSTRUCTORS = gql`
     query EditSessionCategoriesQuery {
@@ -63,59 +53,6 @@ const GET_INSTRUCTORS = gql`
     }
 `;
 
-const GET_SESSION = gql`
-    query EditSessionViewQuery($sessionId: ID!) {
-        session(sessionId: $sessionId) {
-            id
-            isConfirmed
-            startDatetime
-            title
-            instructor {
-                user {
-                    id
-                    firstName
-                    lastName
-                }
-            }
-            course {
-                id
-                isConfirmed
-                title
-                availabilityList {
-                    startTime
-                    endTime
-                    dayOfWeek
-                }
-                room
-                courseCategory {
-                    id
-                    name
-                }
-                instructor {
-                    user {
-                        id
-                        firstName
-                        lastName
-                    }
-                    subjects {
-                        name
-                    }
-                }
-                enrollmentSet {
-                    student {
-                        user {
-                            id
-                            firstName
-                            lastName
-                        }
-                    }
-                }
-            }
-            endDatetime
-            startDatetime
-        }
-    }
-`;
 
 const UPDATE_COURSE = gql`
     mutation UpdateCourse(
